frontend/requirements.txt
gunicorn==19.9.0
pika==0.12.0
python-engineio==2.2.0
python-socketio==2.0.0
six==1.11.0
tornado==5.1
